requirements.txt
amqp==5.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
autopep8==2.3.1; python_version >= '3.8'
billiard==4.2.0; python_version >= '3.7'
blinker==1.8.2; python_version >= '3.8'
brotli==1.1.0
celery==5.4.0; python_version >= '3.8'
certifi==2024.6.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
configargparse==1.7; python_version >= '3.5'
cryptography==42.0.8; python_version >= '3.7'
defusedxml==0.8.0rc2; python_version >= '3.6'
distlib==0.3.8
dj-database-url==2.2.0
django==5.0.6; python_version >= '3.10'
django-cors-headers==4.4.0; python_version >= '3.8'
django-debug-toolbar==4.4.2; python_version >= '3.8'
django-filter==24.2; python_version >= '3.8'
django-redis==5.4.0; python_version >= '3.6'
django-silk==5.1.0; python_version >= '3.8'
django-templated-mail==1.1.1
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djoser==2.2.3; python_version >= '3.8' and python_version < '4.0'
drf-nested-routers==0.94.1; python_version >= '3.8'
filelock==3.15.4; python_version >= '3.8'
flask==3.0.3; python_version >= '3.8'
flask-cors==4.0.1
flask-login==0.6.3; python_version >= '3.7'
flower==2.0.1; python_version >= '3.7'
fsspec==2024.6.1; python_version >= '3.8'
gevent==24.2.1; python_version >= '3.8'
geventhttpclient==2.3.1; python_version >= '3.9'
gprof2dot==2024.6.6; python_version >= '3.8'
greenlet==3.0.3; python_version >= '3.11' and platform_python_implementation == 'CPython'
gunicorn==22.0.0; python_version >= '3.7'
humanize==4.9.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
kombu==5.3.7; python_version >= '3.8'
locust==2.29.1; python_version >= '3.9'
markupsafe==2.1.5; python_version >= '3.7'
model-bakery==1.18.2; python_version >= '3.8'
mpmath==1.3.0
msgpack==1.0.8; python_version >= '3.8'
networkx==3.3; python_version >= '3.10'
numpy==2.0.0; python_version >= '3.12'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
pbr==6.0.0; python_version >= '2.6'
pillow==10.4.0; python_version >= '3.8'
platformdirs==4.2.2; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
prometheus-client==0.20.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
psutil==6.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
psycopg2-binary==2.9.9; python_version >= '3.7'
pycodestyle==2.12.0; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pytest==8.2.2; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz==2024.1
pyzmq==26.0.3; python_version >= '3.7'
redis==5.0.7; python_version >= '3.7'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
setuptools==70.2.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.4.1; python_version >= '3.8'
social-auth-core==4.5.4; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
stevedore==5.2.0; python_version >= '3.8'
sympy==1.12.1; python_version >= '3.8'
torch==2.3.1; python_full_version >= '3.8.0'
torchvision==0.18.1; python_version >= '3.8'
tornado==6.4.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
urllib3==2.2.2; python_version >= '3.8'
vine==5.1.0; python_version >= '3.6'
virtualenv==20.26.3; python_version >= '3.7'
virtualenv-clone==0.5.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
virtualenvwrapper==6.1.0; python_version >= '3.8'
wcwidth==0.2.13
werkzeug==3.0.3; python_version >= '3.8'
whitenoise[brotli]==6.7.0; python_version >= '3.8'
zope.event==5.0; python_version >= '3.7'
zope.interface==6.4.post2; python_version >= '3.7'

django~=5.0.6
django-debug-toolbar~=4.4.2
djoser~=2.2.3
djangorestframework~=3.15.2
openai~=1.38.0
transformers~=4.43.3
torch~=2.3.1
pydantic~=2.8.2
pyyaml~=6.0.1
jinja2~=3.1.4
setuptools~=70.2.0
requests~=2.32.3
django-filter~=24.2